Move duplicate objects in both Video Players to a shared file
